+++ client/src/ui/elements/LoadingAnimation.tsx
@@ -1,0 +1,7 @@
+export const LoadingAnimation = () => {
+  return (
+    <div className="p-10 w-full flex justify-center">
+      <img src="/images/eternum-logo_animated.png" className=" invert scale-50 self-center" />
+    </div>
+  );
+};
